huge tmpfs: fix Mlocked meminfo, track huge & unhuge mlocks
Up to this point, the huge tmpfs effort has barely looked at or touched
mm/mlock.c at all: just a PageTeam test to stop __munlock_pagevec_fill()
crashing (or hanging on a non-existent spinlock) on hugepage pmds.

/proc/meminfo's Mlocked count has been whatever happens to be shown if we
do nothing extra: a hugely mapped and mlocked team page would count as 4kB
instead of the 2MB you'd expect; or at least until the previous
(Unevictable) patch, which now requires lruvec locking for
hpage_nr_pages() on a team page (locking not given it in mlock.c), and
varies the amount returned by hpage_nr_pages().

It would be easy to correct the 4kB or variable amount to 2MB by using an
alternative to hpage_nr_pages() here.  And it would be fairly easy to
maintain an entirely independent PmdMlocked count, such that
Mlocked+PmdMlocked might amount to (almost) twice RAM size.  But is that
what observers of Mlocked want?  Probably not.

So we need a huge pmd mlock to count as 2MB, but discount 4kB for each
page within it that is already mlocked by pte somewhere, in this or
another process; and a small pte mlock to count usually as 4kB, but 0 if
the team head is already mlocked by pmd somewhere.

Can this be done by maintaining extra counts per team?  I did intend so,
but (a) space in team_usage is limited, and (b) mlock and munlock already
involve slow LRU switching, so might as well keep 4kB and 2MB in synch
manually; but most significantly (c): the trylocking around which mlock
was designed, makes it hard to work out just when a count does need to be
incremented.

The hard-won solution looks much simpler than I thought possible, but an
odd interface in its current implementation.  Not so much needed changing,
mainly just clear_page_mlock(), mlock_vma_page() munlock_vma_page() and
try_to_"unmap"_one().  The big difference from before, is that a team head
page might be being mlocked as a 4kB page or as a 2MB page, and the called
functions cannot tell: so now need an nr_pages argument.  But odd because
the PageTeam case immediately converts that to an iteration count, whereas
the anon THP case keeps it as the weight for a single iteration (and in
the munlock case has to reconfirm it under lruvec lock).  Not very nice,
but will do for now: it was so hard to get here, I'm very reluctant to
pull it apart in a hurry.

The TEAM_PMD_MLOCKED flag in team_usage does not play a large part, just
optimizes out the overhead in a couple of cases: we don't want to make yet
another pass down the team, whenever a team is last unmapped, just to
handle the unlikely mlocked-then-truncated case; and we don't want
munlocking one of many parallel huge mlocks to check every page.

Signed-off-by: Hugh Dickins <hughd@google.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Andres Lagar-Cavilla <andreslc@google.com>
Cc: Yang Shi <yang.shi@linaro.org>
Cc: Ning Qu <quning@gmail.com>
Cc: David Rientjes <rientjes@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>