huge tmpfs: fix Mapped meminfo, track huge & unhuge mappings
Maintaining Mlocked was the difficult one, but now that it is correctly
tracked, without duplication between the 4kB and 2MB amounts, I think we
have to make a similar effort with Mapped.

But whereas mlock and munlock were already rare and slow operations, to
which we could fairly add a little more overhead in the huge tmpfs case,
ordinary mmap is not something we want to slow down further, relative to
hugetlbfs.

In the Mapped case, I think we can take small or misaligned mmaps of huge
tmpfs files as the exceptional operation, and add a little more overhead
to those, by maintaining another count for them in the head; and by
keeping both hugely and unhugely mapped counts in the one long, can rely
on cmpxchg to manage their racing transitions atomically.

That's good on 64-bit, but there are not enough free bits in a 32-bit
atomic_long_t team_usage to support this: I think we should continue to
permit huge tmpfs on 32-bit, but accept that Mapped may be doubly counted
there.  (A more serious problem on 32-bit is that it would, I think, be
possible to overflow the huge mapping counter: protection against that
will need to be added.)

Now that we are maintaining NR_FILE_MAPPED correctly for huge tmpfs,
adjust vmscan's zone_unmapped_file_pages() to exclude NR_SHMEM_PMDMAPPED,
which it clearly would not want included.  Whereas minimum_image_size() in
kernel/power/snapshot.c?  I have not grasped the basis for that
calculation, so leaving untouched.

Signed-off-by: Hugh Dickins <hughd@google.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Andres Lagar-Cavilla <andreslc@google.com>
Cc: Yang Shi <yang.shi@linaro.org>
Cc: Ning Qu <quning@gmail.com>
Cc: David Rientjes <rientjes@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>