huge tmpfs: use Unevictable lru with variable hpage_nr_pages
A big advantage of huge tmpfs over hugetlbfs is that its pages can be
swapped out; but too often it OOMs before swapping them out.

At first I tried changing page_evictable(), to treat all tail pages of a
hugely mapped team as unevictable: the anon LRUs were otherwise swamped by
pages that could not be freed before the head.

That worked quite well, some of the time, but has some drawbacks.

Most obviously, /proc/meminfo is liable to show 511/512ths of all the
ShmemPmdMapped as Unevictable; which is rather sad for a feature intended
to improve on hugetlbfs by letting the pages be swappable.

But more seriously, although it is helpful to have those tails out of the
way on the Unevictable list, page reclaim can very easily come to a point
where all the team heads to be freed are on the Active list, but the
Inactive is large enough that !inactive_anon_is_low(), so the Active is
never scanned to unmap those heads to release all the tails.  Eventually
we OOM.

Perhaps that could be dealt with by hacking inactive_anon_is_low(): but it
wouldn't help the Unevictable numbers, and has never been necessary for
anon THP.  How does anon THP avoid this?  It doesn't put tails on the LRU
at all, so doesn't then need to shift them to Unevictable; but there would
still be the danger of an Active list full of heads, holding the unseen
tails, but the ratio too high for for Active scanning - except that
hpage_nr_pages() weights each THP head by the number of small pages the
huge page holds, instead of the usual 1, and that is what keeps the
Active/Inactive balance working.

So in this patch we try to do the same for huge tmpfs pages.  However, a
team is not one huge compound page, but a collection of independent pages,
and the fair and lazy way to accomplish this seems to be to transfer each
tail's weight to head at the time when shmem_writepage() has been asked to
evict the page, but refuses because the head has not yet been evicted.  So
although the failed-to-be-evicted tails are moved to the Unevictable LRU,
each counts for 0kB in the Unevictable amount, its 4kB going to the head
in the Active(anon) or Inactive(anon) amount.

With a few exceptions, hpage_nr_pages() is now only called on a
maybe-PageTeam page while under lruvec lock: and we do need to hold lruvec
lock when transferring weight from one page to another.  Exceptions:
mlock.c (next patch), subsequently self-correcting calls to
page_evictable(), and the "nr_rotated +=" line in shrink_active_list(),
which has no need to be precise.

(Aside: quite a few of our calls to hpage_nr_pages() are no more than ways
to side-step the THP-off BUILD_BUG_ON() buried in HPAGE_PMD_NR: we might
do better to kill that BUILD_BUG_ON() at last.)

Lru lock is a new overhead, which shmem_disband_hugehead() prefers to
avoid, if the head's weight is just the default 1.  And it's not clear how
well this will all play out if different pages of a team are charged to
different memcgs: but the code allows for that, and it should be fine
while that's just an exceptional minority case.

A change I like in principle, but have not made, and do not intend to make
unless we see a workload that demands it: it would be natural for
mark_page_accessed() to retrieve such a 0-weight page from the Unevictable
LRU, assigning it weight again and giving it a new life on the Active and
Inactive LRUs.  As it is, I'm hoping PageReferenced gives a good enough
hint as to whether a page should be retained, when
shmem_evictify_hugetails() brings it back from Unevictable to Inactive.

Signed-off-by: Hugh Dickins <hughd@google.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Andres Lagar-Cavilla <andreslc@google.com>
Cc: Yang Shi <yang.shi@linaro.org>
Cc: Ning Qu <quning@gmail.com>
Cc: David Rientjes <rientjes@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>