huge tmpfs: mem_cgroup move charge on shmem huge pages
Early on, for simplicity, we disabled moving huge tmpfs pages from one
memcg to another (nowadays only required when moving a task into a memcg
having move_charge_at_immigrate exceptionally set).  We're about to add a
couple of memcg stats for huge tmpfs, and will need to confront how to
handle moving those stats, so better enable moving the pages now.

Although they're discovered by the pmd's get_mctgt_type_thp(), they have
to be considered page by page, in what's usually the pte scan: because
although the common case is for each member of the team to be owned by the
same memcg, nowhere is that enforced - perhaps one day we shall need to
enforce such a limitation, but not so far.

Signed-off-by: Hugh Dickins <hughd@google.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Andres Lagar-Cavilla <andreslc@google.com>
Cc: Yang Shi <yang.shi@linaro.org>
Cc: Ning Qu <quning@gmail.com>
Cc: David Rientjes <rientjes@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>