init/main.c: simplify initcall_blacklisted()
Using kasprintf to get the function name makes us look up the name twice,
along with all the vsnprintf overhead of parsing the format string etc. 
It also means there is an allocation failure case to deal with.  Since
symbol_string in vsprintf.c would anyway allocate an array of size
KSYM_SYMBOL_LEN on the stack, that might as well be done up here.

Moreover, since this is a debug feature and the blacklisted_initcalls list
is usually empty, we might as well test that and thus avoid looking up the
symbol name even once in the common case.

Signed-off-by: Rasmus Villemoes <linux@rasmusvillemoes.dk>
Acked-by: Rusty Russell <rusty@rustcorp.com.au>
Acked-by: Prarit Bhargava <prarit@redhat.com>
Cc: Oleg Nesterov <oleg@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>