mm, page_alloc: remove redundant check for empty zonelist
A check is made for an empty zonelist early in the page allocator fast
path but it's unnecessary.  When get_page_from_freelist() is called, it'll
return NULL immediately.  Removing the first check is slower for machines
with memoryless nodes but that is a corner case that can live with the
overhead.

Signed-off-by: Mel Gorman <mgorman@techsingularity.net>
Cc: Vlastimil Babka <vbabka@suse.cz>
Cc: Jesper Dangaard Brouer <brouer@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>