mm/zpool: use workqueue for zpool_destroy
Add a work_struct to struct zpool, and change zpool_destroy_pool to defer
calling the pool implementation destroy.

The zsmalloc pool destroy function, which is one of the zpool
implementations, may sleep during destruction of the pool.  However zswap,
which uses zpool, may call zpool_destroy_pool from atomic context.  So we
need to defer the call to the zpool implementation to destroy the pool.

This is essentially the same as Yu Zhao's proposed patch to zsmalloc,
but moved to zpool.

Signed-off-by: Dan Streetman <ddstreet@ieee.org>
Reported-by: Yu Zhao <yuzhao@google.com>
Cc: Dan Streetman <dan.streetman@canonical.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>